+++ kernels/omp/logarithm.c
@@ -1,7 +1,7 @@
 #include "../../include/kernels.h"
 
 void
-log_libmath (double *input, double *output, int n)
+log_libmath (float *input, float *output, int n)
 {
 #pragma omp for
   for (int i = 0; i < n; i++)
@@ -9,7 +9,7 @@ log_libmath (double *input, double *output, int n)
 }
 
 void
-logarithm_taylor (double *input, double *output, int n)
+logarithm_taylor (float *input, float *output, int n)
 {
   int iter = 100;
 #pragma omp for
@@ -22,8 +22,8 @@ logarithm_taylor (double *input, double *output, int n)
           continue; // Passe à la valeur suivante si x[j] est négatif ou nul
         }
 
-      double result = 0.0;
-      double term = input[j] - 1.0; // Premier terme de la série de Taylor
+      float result = 0.0;
+      float term = input[j] - 1.0; // Premier terme de la série de Taylor
 
       for (int i = 1; i <= iter; i++)
         {
